Modelo page updated

+++ src/pages/features/modelo/modelo-home/modelo-home.ts
@@ -6,14 +6,17 @@ import {
     LoadingController,
     NavController,
     NavParams,
-    ToastController
+    ToastController,
+    ModalController
 } from 'ionic-angular';
+import {DialogoProvider} from "../../../../injectables/dialogo";
 import {ModeloService} from '../../../../services/domain/modelo.service';
 import {ModeloDTO} from '../../../../models/modelo.dto';
+import {LoadingProvider} from "../../../../injectables/loading";
 
 @IonicPage()
 @Component({
-    selector: 'page-modelos',
+    selector: 'page-modelo-home',
     templateUrl: 'modelo-home.html',
 })
 export class ModeloHomePage {
@@ -25,10 +28,13 @@ export class ModeloHomePage {
     });
 
     constructor(
+        public dialogo: DialogoProvider,
         public navCtrl: NavController,
+        public loaderProvider: LoadingProvider,
         public navParams: NavParams,
         public modeloService: ModeloService,
         public events: Events,
+        public modalCtrl: ModalController,
         private toastCtrl: ToastController,
         private alertCtrl: AlertController,
         private loadingCtrl: LoadingController
@@ -70,19 +76,30 @@ export class ModeloHomePage {
     }
 
     delete(modelo: ModeloDTO) {
-        this.modeloService.remove(modelo)
-            .subscribe(response => {
-                    this.events.publish('updateScreen');
-                    let alert = this.alertCtrl.create({
-                        title: 'Sucesso',
-                        subTitle: 'Modelo removido com sucesso!',
-                        buttons: ['Continuar']
-                    });
-                    alert.present();
-                },
-                error => {
-                    this.alertaRemoverErro(modelo);
-                });
+        let mensagem: string = 'Você realmente quer apagar esse registro de defeito?';
+        let titulo: string = 'Confirmar Remoção';
+        let alert = this.dialogo.exibirDialogoConfirmacao(mensagem, titulo);
+
+        alert.present();
+
+        alert.onDidDismiss((confirmado) => {
+            if (confirmado) {
+                this.loading = this.loaderProvider.exibirLoadingPadrao("Apagando o modelo.");
+                this.presentLoading(true);
+
+                this.modeloService.remove(modelo)
+                    .subscribe(() => {
+                            this.events.publish('updateScreen');
+                            this.presentLoading(false);
+                            this.dialogo.exibirToast("Modelo apagado com sucesso.");
+                        },
+                        error => {
+                            this.presentLoading(false);
+                            this.dialogo.exibirToast("Não foi possível remover. O modelo está associado em estoque.");
+                            console.log(error);
+                        })
+            }
+        });
     }
 
     alertaRemoverErro(modelo: ModeloDTO) {
@@ -220,6 +237,18 @@ export class ModeloHomePage {
             console.log('Dismissed toast');
         });
         toast.present();
+    }
+
+    public insert(): void {
+        let modalModelo = this.modalCtrl.create('ModeloInsertPage');
+
+        modalModelo.present();
+
+        modalModelo.onDidDismiss(vendido => {
+            if (vendido) {
+                this.updateModelo;
+            }
+        });
     }
 
 }
